+++ CHANGELOG.md
@@ -10,12 +10,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `DecisionTreeRegressor` model
 - `Tree` Transporter
 - Decision Tree chain
-- `decision_tree_chain.py` 
 - `DecisionTreeRegressor` Test
-### Changed
-- Tests config modified
 - DecisionTree params initialized in `pymilo_param`
 - Decision Tree support added `pymilo_func.py`
+### Changed
+- Tests config modified
 
 ## [0.2] - 2023-08-02
 ### Added