+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.3] - 2023-09-27
 ### Added
 - scikit-learn decision tree models
 - `ExtraTreeClassifier` model
@@ -47,6 +48,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `Export` class
 - `Import` class
 
-[Unreleased]: https://github.com/openscilab/pymilo/compare/v0.2...dev
+[Unreleased]: https://github.com/openscilab/pymilo/compare/v0.3...dev
+[0.3]: https://github.com/openscilab/pymilo/compare/v0.2...v0.3
 [0.2]: https://github.com/openscilab/pymilo/compare/v0.1...v0.2
 [0.1]: https://github.com/openscilab/pymilo/compare/e887108...v0.1